+++ apps/web/auth/auth-client.ts
@@ -1,5 +1,9 @@
 import { createAuthClient } from 'better-auth/react';
 
 export const authClient = createAuthClient({
-  baseURL: `${process.env.NEXT_PUBLIC_API_URL}/api/auth`,
+  baseURL: process.env.NEXT_PUBLIC_API_URL,
+  basePath: '/api/auth',
+  fetchOptions: {
+    credentials: 'include',
+  },
 });
