+++ apps/web/components/app-command-k.tsx
@@ -1,0 +1,11 @@
+'use client';
+
+import { useRouter } from 'next/navigation';
+
+import { CommandK } from '@workspace/client/components';
+
+export function AppCommandK() {
+  const router = useRouter();
+
+  return <CommandK onNavigate={(path) => router.push(path)} />;
+}
