--- apps/web/components/document-attachments.tsx
@@ -1,1 +1,0 @@
-export * from '@workspace/client/components/document-attachments';
