--- apps/web/components/document-preview.tsx
@@ -1,1 +1,0 @@
-export * from '@workspace/client/components/document-preview';
