--- apps/word-add-in/src/taskpane/utils/api.ts
@@ -1,13 +1,0 @@
-import {
-  getApiBaseUrl as getSharedApiBaseUrl,
-  resolveApiUrl,
-  setApiBaseUrl,
-} from "@workspace/client/lib/api";
-
-const DEFAULT_BASE_URL = "https://mt-gpt-production.up.railway.app";
-
-setApiBaseUrl(process.env.WORD_ADDIN_API_URL ?? DEFAULT_BASE_URL);
-
-export const getApiBaseUrl = () => getSharedApiBaseUrl();
-
-export const getApiUrl = (path: string) => resolveApiUrl(path);
